[REFACTOR] @CurrentUser 삭제
