+++ src/main/java/com/be1/plant4you/auth/controller/UserController.java
@@ -2,8 +2,6 @@ package com.be1.plant4you.auth.controller;
 
 import com.be1.plant4you.auth.dto.response.UserNameResponse;
 import com.be1.plant4you.auth.service.UserService;
-import com.be1.plant4you.common.auth.CurrentUser;
-import com.be1.plant4you.common.auth.UserPrincipal;
 import io.swagger.annotations.Api;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
@@ -22,14 +20,14 @@ public class UserController {
 
     @Operation(summary = "현재 로그인한 유저 이름 조회", description = "로그인 후 메인화면에 '~님 안녕하세요'에 사용")
     @GetMapping("/name")
-    public UserNameResponse getLoggedInUserName(@CurrentUser UserPrincipal userPrincipal) {
-        return userService.getUserName(userPrincipal.getUserId());
+    public UserNameResponse getLoggedInUserName() {
+        return userService.getUserName();
     }
 
     @Operation(summary = "유저 탈퇴", description = "일단 유저 관련 정보 모두 삭제하도록 구현")
     @DeleteMapping
-    public String withdraw(@CurrentUser UserPrincipal userPrincipal) {
-        userService.withdraw(userPrincipal.getUserId());
+    public String withdraw() {
+        userService.withdraw();
         return "탈퇴가 완료되었습니다.";
     }
 }
