+++ src/main/java/com/be1/plant4you/auth/oauth/service/CustomOAuth2UserService.java
@@ -9,7 +9,7 @@ import com.be1.plant4you.auth.repository.RefreshTokenRepository;
 import com.be1.plant4you.auth.repository.UserRepository;
 import com.be1.plant4you.auth.token.AuthToken;
 import com.be1.plant4you.auth.token.AuthTokenProvider;
-import com.be1.plant4you.common.auth.UserPrincipal;
+import com.be1.plant4you.auth.UserPrincipal;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
