--- src/main/java/com/be1/plant4you/common/auth/CurrentUser.java
@@ -1,12 +1,0 @@
-package com.be1.plant4you.common.auth;
-
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-
-import java.lang.annotation.*;
-
-@Target({ ElementType.PARAMETER, ElementType.ANNOTATION_TYPE })
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@AuthenticationPrincipal
-public @interface CurrentUser {
-}
